requirements.txt
python-dotenv
pdfplumber
openai
pinecone-client
langchain
sentence-transformers
pypdf